fix: correct ValueInput/ValueOutput package to world.level.storage, fix moveTo double signature

+++ src/main/java/arghorror/ArchitectSpawner.java
@@ -66,7 +66,8 @@ public class ArchitectSpawner {
         TheArchitectEntity entity = ModEntities.THE_ARCHITECT.create(level, EntitySpawnReason.MOB_SUMMONED);
         if (entity == null) return;
 
-        entity.moveTo(spawnPos.getX() + 0.5, spawnPos.getY(), spawnPos.getZ() + 0.5, 0f, 0f);
+        // moveTo(double x, double y, double z, float yRot, float xRot)
+        entity.moveTo(spawnPos.getX() + 0.5, (double) spawnPos.getY(), spawnPos.getZ() + 0.5, 0f, 0f);
         level.addFreshEntity(entity);
         activeEntities.put(player.getUUID(), entity);
     }
